+++ packages/runtime-webpack/src/index.ts
@@ -1,0 +1,2 @@
+import RxJSDebuggingPlugin from './RxJSDebuggingPlugin';
+export default RxJSDebuggingPlugin;