+++ packages/telemetry/src/operatorIdentifier/serialize.ts
@@ -1,0 +1,10 @@
+import { IOperatorIdentifier } from '.';
+
+export default function serializeOperatorIdentifier({
+  character,
+  fileName,
+  line,
+  operatorIndex,
+}: IOperatorIdentifier): string {
+  return JSON.stringify({ character, fileName, line, operatorIndex });
+}